Fixed a basic Gameloop system
created a system to handle a basic game loop, keeping track of score, and the leader.

Added player tracking of who pushed them recently and the addition of score on death if one is prey

added randomized prey in the gameloopController

Created a new script for player action that handles pushing and trap laying and moved the code from trapability and player into it, making it easy to keep track of and controll.