Implementation of basic gameloop system
Implemented a game looping system.
implemented randomized prey setting (WIP)
implemented score and score tracking
Created a script called PlayerActionsController and moved the code in trapability and player related to pushing or placing traps into it for easier oversight of the actions and to be able to easily swap what actions one can do.
implemented a round reset on positions aswell as a reset on all players position when prey dies (WIP)