Bug creating afterimage has been discovered
A new type of bug has been found that follows the characters persistently and is creating an afterimage, The bug seems to be fulfilling a purpose in life and should remain undisturbed unless it stars to cause trouble. For now all we need to do is welcome our newest bug to the collective!