created a new collision system for players
now the players collision for danger and traps will be on its feet.

will require a rework on the colliders on the map to make it perfect.

but this will make the collision feel more fair.

a small tweak to rotation of said collider needs to be done aswell depending on how you run.