start of death implementation system
the start of a system to handle the different ways for characters to die.

left todo import and make the different death animations and put them in the animation controller

make the animation call on the startghost event instead of the code.

set the right triggers to be triggered inside the deathcontroller code.