made assist and pushed by go null also removed warning
removed warning on the dontdestroyonload on audio originating from the fact that it was not a root object (´cant be child to anything)

made it so that if the hunter dies and has an assist push it goes null on the prey.
made it so that if the hunter dies and has pushed the prey recently isPushed on prey goes null.